feat(validation)!: gate model-loading tests behind `gpu-tests`
Whisper's validation conflated two questions. `download` said "fetch the
assets", and the choice of backend feature was read as "run the expensive
comparisons" — so a plain `--features wgpu` did something the caller had
not asked for, and a `compile_error!` refused builds that were merely
selecting a backend.

`gpu-tests` states the intent directly. It is orthogonal to the backend
features — it selects *whether* the model-loading tests compile, a backend
feature selects *what* they run on — and maps to `[]`, enabling no
dependency.

The module tree now mirrors the gates, so a test's path names what it needs:

    audio/                        fixture integrity, no model
    audio/gpu_tests.rs            needs an accelerator
    audio/download/               needs the fetched assets
    audio/download/gpu_tests.rs   needs both

With `download,gpu-tests` the suite is twelve tests; with either alone it is
the three fixture-integrity checks.

STYLE.md gains a `cargo features` chapter reserving the names this workspace
uses — the backend trio, `gpu-tests`, `download`, `onnx_gen`, `checkpoint` —
so the convention is written down ahead of full adoption. It also widens the
file's stated remit, which claimed to govern in-source rustdoc only.

A `gpu.yml` workflow runs the full set; the ordinary CI run stays cheap.

Four documents described a backend guard that no longer exists, or in
Silero's case never did: its docs claimed the cross-checks "are not
compiled, and a single ignored test says so", while `mod cross_test;` is
unconditional and both run on the CPU fallback. Corrected to say what is
true, and to name `gpu-tests` as the convention Silero has not yet adopted.

BREAKING CHANGE: `cargo test -p whisper-model-validation --features
download,wgpu` no longer runs the model-loading tests; add `gpu-tests`.
