+++ STYLE.md
@@ -1,11 +1,12 @@
 # Bunsen Style Guide
 
-Conventions for code and documentation in the `bunsen` crate. This file is the
-source of truth; assertions added under each chapter are applied across the code
-base.
+Conventions for code, documentation, and build configuration across the
+`bunsen` workspace. This file is the source of truth; assertions added under
+each chapter are applied across the code base.
 
 > See also [`book/src/contributing/style.md`](book/src/contributing/style.md)
-> for prose/Book conventions. This file governs in-source `rustdoc`.
+> for prose/Book conventions. This file governs in-source `rustdoc` and the
+> `Cargo.toml` manifests.
 
 ## rustdoc
 
@@ -64,5 +65,109 @@ These are **not** tensor shapes — leave them as written:
 - Rust type code spans: `&[usize; D]`, `Param<Tensor<B, R, K>>`.
 - Half-open ranges and indexing: `[start, end)`, `env[$VAR]`.
 - Intra-doc link syntax: `[text](url)`.
+
+## cargo features
+
+### Conventional names
+
+* `wgpu`, `cuda`, `metal` — select an accelerator backend.
+* `gpu-tests` — compile the tests that need one.
+* `download` — the build may reach the network.
+* `onnx_gen` — generate reference models from an ONNX graph.
+* `checkpoint` — fetch pretrained weights.
+
+These names are **reserved** across the workspace. No crate is obliged to
+offer one; a crate that offers one means this, and nothing else. Most crates
+predate the list, so this is an interim expectation — new features are named
+*to* it rather than around it.
+
+### Backend selection
+
+`cuda`, `metal` and `wgpu` each select one accelerator, and are **never** in
+`default`. Precedence, when more than one is set, is `cuda` > `metal` >
+`wgpu` > `flex`.
+
+A backend feature enables the backend in **`bunsen`**, not only in `burn`:
+
+```text
+prefer:  wgpu = ["bunsen/wgpu"]                (a test picks the backend)
+         wgpu = ["burn/wgpu", "bunsen/wgpu"]   (and `burn` is used directly)
+not:     wgpu = ["burn/wgpu"]                  (`PerformanceBackend` is Flex)
+```
+
+Only `bunsen/<backend>` moves `bunsen::support::testing::PerformanceBackend`.
+Its `cfg_select!` falls through to `Flex`, a CPU backend, when no backend
+feature reaches `bunsen` — so a test written against `PerformanceBackend`
+still compiles and still passes, having quietly measured the CPU.
+
+`flex` names that fallback rather than an accelerator, and is exempt from the
+prohibition above: an example that wants to run without hardware carries it in
+`default`.
+
+### GPU tests
+
+`gpu-tests` gates tests too slow to run without an accelerator. It is off by
+default, and maps to `[]` — it enables no dependency and pulls in no backend.
+It selects only which tests compile.
+
+It is **orthogonal** to the backend features: `gpu-tests` selects *whether*
+the expensive tests are built, a backend feature selects *which* accelerator
+runs them. Both are passed:
+
+```text
+cargo test --release -p whisper-model-validation --features gpu-tests,wgpu
+```
+
+Do **not** infer the intent from a backend feature:
+
+```text
+prefer:  #[cfg(all(test, feature = "gpu-tests"))]
+not:     #[cfg(any(feature = "wgpu", feature = "cuda", feature = "metal"))]
+```
+
+The second conflates *an accelerator is available* with *the slow tests were
+asked for*, and leaves no way to have one without the other. A
+`compile_error!` on the same condition is the same conflation, louder.
+
+### Network access
+
+`download` marks a build that may reach the network. It is **off by default**
+— a plain `cargo build --workspace` never fetches. Assets are pinned to a
+digest, re-verified on every build, and cached outside `OUT_DIR`, so
+`cargo clean` does not force a re-download. An environment variable points the
+build at a local file instead.
+
+`bunsen-bundled-whisper` is the one exception, enumerated rather than implied:
+`checkpoint` is in its `default`, so a workspace build does fetch 145 MB on a
+cold cache. Naming the exception keeps the rule absolute everywhere else.
+
+### Model assets
+
+`onnx_gen` generates Rust reference models from an ONNX graph and exposes them
+in a `crate::onnx_gen` module. `checkpoint` fetches pretrained weights.
+
+Neither implies the other — they are different assets. `download` is the
+aggregate switch a consumer flips; `onnx_gen` and `checkpoint` are the
+per-asset switches on the crate that owns them.
+
+### Feature documentation
+
+Every feature carries a `##` doc comment, in the form `document-features`
+renders:
+
+```text
+## Fetch the pretrained checkpoint and expose it as [`base_pt`].
+checkpoint = []
+```
+
+A crate whose features are part of its public surface renders them into its
+crate docs:
+
+```text
+#![doc = document_features::document_features!()]
+```
+
+The comment is required either way — the manifest is the first place a reader
+looks.
 
 <!-- Assertions go here. -->
