+++ crates/dev/README.md
@@ -1,0 +1,3 @@
+# Development Crates
+
+This directory contains development and testing crates that are not published to crates.io.
