+++ crates/public/README.md
@@ -1,0 +1,3 @@
+# Public Crates
+
+This directory contains public crates that are published to crates.io.
