+++ crates/validation/whisper-model-validation/src/audio/download/gpu_tests.rs
@@ -1,0 +1,283 @@
+use bunsen::{
+    kits::speech::whisper::GreedyDecodeConfig,
+    prelude::TensorElemOpExt,
+    support::testing::{
+        PerformanceBackend,
+        asr::text_error_rate,
+    },
+};
+use burn::{
+    Tensor,
+    prelude::{
+        Backend,
+        Device,
+        Int,
+        TensorData,
+    },
+    tensor::{
+        Tolerance,
+        backend::BackendTypes,
+    },
+};
+
+use crate::{
+    N_FRAMES,
+    N_MELS,
+    audio::{
+        FIXTURES,
+        Reference,
+        bunsen_model,
+        clip_mels,
+        report,
+        to_text,
+        transcript,
+        vocab,
+    },
+    reference,
+};
+
+type B = PerformanceBackend;
+type F = <B as BackendTypes>::FloatElem;
+
+/// `<|startoftranscript|> <|en|> <|transcribe|> <|notimestamps|>`.
+pub const PROMPT: [i64; 4] = [50258, 50259, 50359, 50363];
+/// `<|endoftext|>`.
+pub const EOT: i64 = 50257;
+
+/// Splits log-mels into the encoder's fixed windows, zero-padding the
+/// last.
+fn windows_of<B: Backend>(
+    mels: Tensor<B, 3>,
+    device: &Device<B>,
+) -> Vec<Tensor<B, 3>> {
+    let frames = mels.dims()[2];
+    (0..frames.div_ceil(N_FRAMES))
+        .map(|w| {
+            let start = (w * N_FRAMES) as isize;
+            let end = ((w + 1) * N_FRAMES).min(frames) as isize;
+            let win = mels.clone().slice_dim(2, start..end);
+
+            let have = win.dims()[2];
+            if have < N_FRAMES {
+                Tensor::cat(
+                    vec![win, Tensor::zeros([1, N_MELS, N_FRAMES - have], device)],
+                    2,
+                )
+            } else {
+                win
+            }
+        })
+        .collect()
+}
+
+/// Greedily decodes one window against the reference decoder.
+///
+/// The export is KV-cache-free, so the whole prefix is re-fed every
+/// step — which is slow, and exactly why bunsen has a cache.
+fn greedy_reference<B: Backend>(
+    decoder: &reference::decoder::Model<B>,
+    xa: Tensor<B, 3>,
+    device: &Device<B>,
+) -> Vec<i64> {
+    let mut prefix = PROMPT.to_vec();
+    let mut out = Vec::new();
+
+    for _ in 0..224 {
+        let len = prefix.len();
+        let tokens: Tensor<B, 2, Int> =
+            Tensor::from_data(TensorData::new(prefix.clone(), [1, len]), device);
+
+        let picked: Vec<i64> = decoder
+            .forward(tokens, xa.clone())
+            .0
+            .slice_dim(1, (len - 1) as isize..len as isize)
+            .argmax(2)
+            .into_data()
+            .convert::<i64>()
+            .to_vec()
+            .unwrap();
+
+        if picked[0] == EOT {
+            break;
+        }
+        out.push(picked[0]);
+        prefix.push(picked[0]);
+    }
+
+    out
+}
+
+/// **The agreement gate.** bunsen must decode what `openai-whisper`
+/// decodes, on the same audio and the same windowing.
+#[test]
+fn test_bunsen_agrees_with_openai_reference() {
+    let device: Device<B> = Default::default();
+    let table = vocab();
+    let model = bunsen_model::<B>(&device);
+    let config = GreedyDecodeConfig::new(PROMPT.to_vec(), EOT);
+
+    for fixture in FIXTURES {
+        let mels = clip_mels(fixture.name, &device);
+        let got = to_text(&table, &model.decode_chunked(mels, &config));
+        let want = Reference::load(fixture.name).text();
+        let wer = text_error_rate(&got, &want);
+
+        eprintln!("{}: bunsen WER vs openai-whisper {wer:.4}", fixture.name);
+        assert!(
+            wer <= fixture.max_reference_wer,
+            "{}",
+            report(
+                "openai-reference",
+                fixture.name,
+                &got,
+                &want,
+                wer,
+                fixture.max_reference_wer,
+            ),
+        );
+    }
+}
+
+/// The encoder comparison, on **real** log-mels rather than synthetic
+/// ones.
+///
+/// Real speech is not the same test: `synthetic_mels` is a bounded
+/// sawtooth, while a log-mel spectrogram has the dynamic range and the
+/// near-silent bins a mis-scaled layer norm would show up in.
+#[test]
+fn test_onnx_encoder_matches_bunsen_on_real_audio() {
+    let device: Device<B> = Default::default();
+    let reference = reference::encoder::Model::<B>::load_pretrained(&device);
+    let ours = bunsen_model(&device);
+
+    for fixture in FIXTURES {
+        let mels = clip_mels(fixture.name, &device);
+        for (w, window) in windows_of(mels, &device).into_iter().enumerate() {
+            let theirs = reference.forward(window.clone());
+            let mine = ours.forward_encoder(window);
+
+            assert_eq!(mine.dims(), theirs.dims(), "{} window {w}", fixture.name);
+            mine.to_data_as::<F>()
+                .assert_approx_eq::<F>(&theirs.to_data_as::<F>(), Tolerance::rel_abs(1e-1, 2e-2));
+        }
+    }
+}
+
+/// **The ONNX reference transcribes the clip**, independent of bunsen:
+/// the ONNX encoder feeds the ONNX decoder.
+#[test]
+fn test_onnx_reference_transcribes_real_audio() {
+    let device: Device<B> = Default::default();
+    let table = vocab();
+    let reference_enc = reference::encoder::Model::<B>::load_pretrained(&device);
+    let reference_dec = reference::decoder::Model::<B>::load_pretrained(&device);
+
+    for fixture in FIXTURES {
+        let mels = clip_mels(fixture.name, &device);
+        let ids: Vec<Vec<i64>> = windows_of(mels, &device)
+            .into_iter()
+            .map(|window| {
+                let xa = reference_enc.forward(window);
+                greedy_reference(&reference_dec, xa, &device)
+            })
+            .collect();
+
+        let got = to_text(&table, &ids);
+        let want = transcript(fixture.name);
+        let wer = text_error_rate(&got, &want);
+
+        eprintln!("{}: onnx WER vs transcript {wer:.4}", fixture.name);
+        assert!(
+            wer <= fixture.max_wer,
+            "{}",
+            report(
+                "onnx-transcript",
+                fixture.name,
+                &got,
+                &want,
+                wer,
+                fixture.max_wer
+            ),
+        );
+    }
+}
+
+/// **The agreement gate, end to end.** bunsen and the ONNX reference
+/// must transcribe the same clip to the same words.
+///
+/// This is the one that catches what the staged comparisons let
+/// through: each stage can agree inside tolerance while the
+/// composition diverges, because a greedy argmax turns a small
+/// numerical difference into a different word.
+#[test]
+fn test_onnx_reference_and_bunsen_transcribe_alike() {
+    let device: Device<B> = Default::default();
+    let table = vocab();
+    let reference_enc = reference::encoder::Model::<B>::load_pretrained(&device);
+    let reference_dec = reference::decoder::Model::<B>::load_pretrained(&device);
+    let ours = bunsen_model(&device);
+    let config = GreedyDecodeConfig::new(PROMPT.to_vec(), EOT);
+
+    for fixture in FIXTURES {
+        let mels = clip_mels(fixture.name, &device);
+
+        let mut theirs = Vec::new();
+        let mut mine = Vec::new();
+        for window in windows_of(mels, &device) {
+            let xa = reference_enc.forward(window.clone());
+            theirs.push(greedy_reference(&reference_dec, xa, &device));
+            mine.push(ours.decode_window(window, &config));
+        }
+
+        let (want, got) = (to_text(&table, &theirs), to_text(&table, &mine));
+        let wer = text_error_rate(&got, &want);
+
+        eprintln!("{}: bunsen WER vs onnx reference {wer:.4}", fixture.name);
+        assert!(
+            wer <= fixture.max_reference_wer,
+            "{}",
+            report(
+                "onnx-reference",
+                fixture.name,
+                &got,
+                &want,
+                wer,
+                fixture.max_reference_wer,
+            ),
+        );
+    }
+}
+
+/// **The accuracy gate.** Real audio, real weights, judged against the
+/// ground-truth transcript.
+#[test]
+fn test_bunsen_accuracy_against_transcript() {
+    let device: Device<B> = Default::default();
+    let table = vocab();
+    let model = bunsen_model::<B>(&device);
+    let config = GreedyDecodeConfig::new(PROMPT.to_vec(), EOT);
+
+    for fixture in FIXTURES {
+        let mels = clip_mels(fixture.name, &device);
+        let per_window = model.decode_chunked(mels, &config);
+        assert_eq!(per_window.len(), fixture.windows, "{}", fixture.name);
+
+        let got = to_text(&table, &per_window);
+        let want = transcript(fixture.name);
+        let wer = text_error_rate(&got, &want);
+
+        eprintln!("{}: bunsen WER vs transcript {wer:.4}", fixture.name);
+        assert!(
+            wer <= fixture.max_wer,
+            "{}",
+            report(
+                "transcript",
+                fixture.name,
+                &got,
+                &want,
+                wer,
+                fixture.max_wer
+            ),
+        );
+    }
+}
