+++ crates/validation/whisper-model-validation/src/audio/download/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "gpu-tests")]
+mod gpu_tests;
